Reworked frequency and power setting
Frequency can now only be in Hz, and power can only be in dBm. Both
are stored internally in `float` format. Both can still be set through
their respective setters, but the onus is on the user to make sure
they have valid values, in Hz and dBm respectively.
Future work would include the use of the `pint` module and support
for other frequency units.
Additionally, finished writing docstrings for this module.