Small tweaks after lab testing
Made sure the signal generator functionality works well. Might
have missed some corner cases, but the basics work.
There was a weird bug where the `inst0` part of the VISA address
has to be lowercase for the N9000A. Specs say it should be
case-insensitive but hey ho. Other changes include making the
VISA connection a private variable, and returning the spectrum
analyser to a continuous mode after the initial reset.